+++ frontend/src/routes/AuthenticationRoutes.jsx
@@ -1,12 +1,10 @@
 import { lazy } from 'react';
 
-// project imports
-import Loadable from 'ui-component/Loadable';
-import MinimalLayout from 'layout/MinimalLayout';
+import Loadable from 'components/Loadable';
+import MinimalLayout from 'layouts/MinimalLayout';
 
-// maintenance routing
-const LoginPage = Loadable(lazy(() => import('views/pages/authentication/Login')));
-const RegisterPage = Loadable(lazy(() => import('views/pages/authentication/Register')));
+const LoginPage = Loadable(lazy(() => import('pages/authentication/Login')));
+const RegisterPage = Loadable(lazy(() => import('pages/authentication/Register')));
 
 // ==============================|| AUTHENTICATION ROUTING ||============================== //
 
@@ -14,14 +12,8 @@ const AuthenticationRoutes = {
   path: '/',
   element: <MinimalLayout />,
   children: [
-    {
-      path: '/pages/login',
-      element: <LoginPage />
-    },
-    {
-      path: '/pages/register',
-      element: <RegisterPage />
-    }
+    { path: '/login', element: <LoginPage /> },
+    { path: '/register', element: <RegisterPage /> }
   ]
 };